+++ corehq/blobs/tests/test_migrate.py
@@ -8,6 +8,7 @@ from corehq.blobs.mixin import BlobMixin
 from corehq.blobs.s3db import maybe_not_found
 from corehq.blobs.tests.util import (TemporaryFilesystemBlobDB,
     TemporaryMigratingBlobDB, TemporaryS3BlobDB)
+from corehq.util.couch_doc_processor import ResumableDocsByTypeIterator, DOC_PROCESSOR_ITERATION_KEY_PREFIX
 from corehq.util.test_utils import trap_extra_setup
 
 from django.conf import settings
@@ -51,10 +52,10 @@ class BaseMigrationTest(TestCase):
     @staticmethod
     def discard_migration_state(slug):
         doc_types = mod.MIGRATIONS[slug].doc_type_map
-        mod.ResumableDocsByTypeIterator(
+        ResumableDocsByTypeIterator(
             list(doc_types.values())[0].get_db(),
             doc_types,
-            slug + "-blob-migration",
+            slug + DOC_PROCESSOR_ITERATION_KEY_PREFIX,
         ).discard_state()
         mod.BlobMigrationState.objects.filter(slug=slug).delete()
